Spec #20 Finished
lesson/course_id deleted and added into lesson_category/course_id
