+++ src/main/java/MainDemo.java
@@ -56,21 +56,21 @@ public class MainDemo {
 			session.save(course3);
 
 			// LessonCategory
-			LessonCategory lessonCategory1 = new LessonCategory("Comida");
+			LessonCategory lessonCategory1 = new LessonCategory("Comida", course1);
 			session.save(lessonCategory1);
 
-			LessonCategory lessonCategory2 = new LessonCategory("Vehicles");
+			LessonCategory lessonCategory2 = new LessonCategory("Vehicles", course2);
 			session.save(lessonCategory2);
 
 			// Lesson
-			Lesson lesson1 = new Lesson("Comida 1", course1, lessonCategory1);
-			session.save(lesson1);
-
-			Lesson lesson2 = new Lesson("Comida 2",course2, lessonCategory1);
-			session.save(lesson2);
-
-			Lesson lesson3 = new Lesson("Food 3",course3, lessonCategory2);
-			session.save(lesson3);
+//			Lesson lesson1 = new Lesson("Comida 1", lessonCategory1);
+//			session.save(lesson1);
+//
+//			Lesson lesson2 = new Lesson("Comida 2", lessonCategory1);
+//			session.save(lesson2);
+//
+//			Lesson lesson3 = new Lesson("Food 3", lessonCategory2);
+//			session.save(lesson3);
 
 			// Final commit
 			t.commit();
