+++ src/main/java/ams2/linguo/interfaces/ILessonCategoryQueries.java
@@ -1,0 +1,11 @@
+package ams2.linguo.interfaces;
+
+import java.util.List;
+
+import ams2.linguo.model.Course;
+import ams2.linguo.model.LessonCategory;
+
+public interface ILessonCategoryQueries {
+	public List<LessonCategory> getLessonsCategoryByCourseId(long courseId);
+	public LessonCategory insertLessonCategoryByTitle(String title, Course course);
+}
